Upgrade TDLib to tdlib/td@369ee92 + Gradle to `9.1.0` + Reworked biometrics authentication + Build system updates and other changes

+++ app/src/main/java/org/thunderdog/challegram/charts/BaseChartView.java
@@ -363,7 +363,7 @@ public abstract class BaseChartView<T extends ChartData, L extends LineViewData>
     private void measureHeightThreshold() {
         int chartHeight = getMeasuredHeight() - chartBottom;
         if (animateToMaxHeight == 0 || chartHeight == 0) return;
-        thresholdMaxHeight = ((float) animateToMaxHeight / chartHeight) * SIGNATURE_TEXT_SIZE;
+        thresholdMaxHeight = (animateToMaxHeight / chartHeight) * SIGNATURE_TEXT_SIZE;
     }
 
 
@@ -562,7 +562,7 @@ public abstract class BaseChartView<T extends ChartData, L extends LineViewData>
 
         float additionalOutAlpha = 1f;
         if (n > 2) {
-            float v = (a.values[1] - a.values[0]) / (float) (currentMaxHeight - currentMinHeight);
+            float v = (a.values[1] - a.values[0]) / (currentMaxHeight - currentMinHeight);
             if (v < 0.1) {
                 additionalOutAlpha = v / 0.1f;
             }
@@ -590,7 +590,7 @@ public abstract class BaseChartView<T extends ChartData, L extends LineViewData>
 
         float additionalOutAlpha = 1f;
         if (n > 2) {
-            float v = (a.values[1] - a.values[0]) / (float) (currentMaxHeight - currentMinHeight);
+            float v = (a.values[1] - a.values[0]) / (currentMaxHeight - currentMinHeight);
             if (v < 0.1) {
                 additionalOutAlpha = v / 0.1f;
             }
